test: extract MockLaunchAtLoginService.swift from KeychainTestSupport.swift
Splits the `MockLaunchAtLoginService` test double out of
KeychainTestSupport.swift into its own file (the mock is unrelated to
keychain — its home in a keychain-named test-support file was
incidental).

Byte-identical move of lines 42-63. Both files retain their test
imports.

KeychainTestSupport.swift: 63 → 40 lines. Tests: 667.
